layer.conf: Update to whinlatter release series
Drop the styhead and walnascar and update to whinlatter.

Remove all the S = ${WORKDIR}/git assignments works because BB_GIT_DEFAULT_DESTSUFFIX
is set to match S from bitbake.conf (which itself is set to match typical tarball
releases).

A few recipes are setting S to a sub-directory of the git tree and need
to be adjusted accordingly.

Signed-off-by: Sairamreddy Bojja <sbojja@qti.qualcomm.com>